Update 20241027 - Chunking Support
The program recently encountered an issue with rate limiting and was forced to support chunking. This allows multiple track IDs to have their tracks fetched but for fetching their audio features, it only allows 1 id per request and thus, the need to put sleep in it is required.

NOTE: "More Prediction Test" section is still under development. It was delayed due to rate limit incidents.

[CHANGELOG]
🟢 Added the `sleep` import.
🟢 Added the `chunkList` function which separates a list into a list of chunks (2D array).
🟡 Updated .gitignore to ignore the large "tracks_features.csv" file.
🟡 Updated the API functions to properly show the `res` content when 429 is thrown.
🟡 Updated the process of the "More Prediction Test" section to cut the total amount of data from 1204025 to just 1000.
🟡 Updated the process of the "More Prediction Test" section to chunk the IDs into 100 items per chunk before getting the track popularity of each IDs.